Ok, so I gave up fighting the GC
What was happening?
- after clicking the "Hello" button ~30 times, the OCaml GC had apparently
decided that the callback functions for the color changing buttons were
no longer necessary, and cleaned them up.
Clicking one of those buttons resulted in a segfault, or sometimes a bus error.

So now I just pass an ID over to the native side, and maintain a hash
table of id=>callback, ensureing that ocaml knows these callbacks can't
be GC'd.

It's not as pretty imo, but I'm done fighting.